// client/src/app/api/model/fullUserDTO.ts
/**
 * CPOO Server API
 * This is a prototype of CPOO Project\'s front/back API. 
 *
 * The version of the OpenAPI document: 0.0.1
 * Contact: contact@mightycode.fr
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FullUserDTO { 
    login: string;
    password: string;
    /**
     * To choose if the user wants to remember their information for next time
     */
    remember: boolean;
    /**
     * Number (ref) of the profile icon (image) of the user
     */
    icon: number;
    firstname: string;
    lastname: string;
    birthday: string;
    address: string;
}